requirements.txt
git+https://github.com/dk-butsuri/keihan_tracker.git
fastapi
jinja2
uvicorn
python-dotenv
google-genai